Adjust to background changes
The actual background drawing is now delegated to a content, which
is also where the brightness and sharpness properties moved that are
animated during the overview transition. We do not have a content
with those properties, so everything is broken again :-(

For a fix, change how the extension works: Instead of adding the logo to
each backgroundGroup on monitor changes, monkey-patch BackgroundManager
to create the logo together with the backgroundActor, and use that as
parent instead of the group.

Patch adapted from https://pagure.io/background-logo-extension/pull-request/27